requirements.txt
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
Deprecated==1.2.14
Flask==3.0.3
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
pycparser==2.22
PyGithub==2.3.0
PyJWT==2.9.0
PyNaCl==1.5.0
pytest==8.3.2
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
